Skip model validation.
The update_product_details management command doesn't touch any model in
any app. If we want to encourage people to not keep all of the
product_details JSON in their repo (and I think we do) or if they need
to bootstrap at all, and any of their models depend on product_details
data, update_product_details fails.

For example, in Mozillians, the `choices` field for which country you
can list in your profile comes from
`product_details.get_region('en-US')`. Which means trying to run
update_product_details with no `en-US.json` file fails validation, so it
can't run the command.

This just tells Django that it can skip model validation for this
management command.