Fix OSError: No such file or directory
If you import product_details, in __init__.py it goes to build a
ProductDetails instance. If the json_dir doesn't exist, that fails with
an OSError.

That's problematic because then it becomes difficult in some situations
to run "./manage.py update_product_details".

This fixes that so that it checks to see if the directory exists first.

Fixes #34