Fix #42: Add json data storage backends.
Default backend is filesystem that works the way
the current version does. The other built-in backend
is for the Database. Switching among them is a setting
with the dotted path to the backend class.

This also adds to our test coverage and greatly compartmentalizes
and simplifies the code in the ProductDetails class itself.

Also support for Django < 1.7 and Python < 2.7 is dropped.