Fix migration error caused by missing last_modified data
The fix for #72 caused the last_modified data for directories
not to be written, then when the migrations ran there was no
data for those entries and the not-null contstraint failed.
We're not using the last-modified data on directories, but
we should still store it until we do a more thorough overhaul
of the sync mechanism.