Add working config for running tox with Github Actions
This changeset retains the ability to run tox locally (with its breadth of
coverage) while also leveraging GH Actions as the CI runner.

To ensure we had appropriate versions of Python available, we've gone with
the matrix approach at the GHA level, and then taking advantage of the way
that tox will gracefully handle the absence of a particular python version
and keep going.

This does mean that GH tries to run tox for all versions in each python
env, but it's so fast that it should not be a blocker. It does make the
output in CI a bit harder to scan, though

Also, note that Python 3.8 is currently the default version and so is
available in all GHA envs we're using, which means that the 3.8 tests
are run every time along with the specific 3.7, 3.9 or pypy tests. We
could comment it out, but then it leaves a gap if/when GHA moves to 3.9
as the default.