Improve GH Actions configuration + tox setup; drop PyPy tests; support 3.10
This changeset cleans up the matrix strategy used in the GH action to only run
combinations we know are viable. This is more efficient and explicit.

Python 3.10 was added as a test CPython and seems to work fine.

Work towards the above changes also resulted in the PyPy tests failing, so
attempts were made to re-enable these, but it turns out that since Django 2.2,
the version of sqlite bundled inside PyPy has needed upgrading to be compatible
with it (https://docs.djangoproject.com/en/4.0/releases/2.2/#miscellaneous).
For now, PyPy tests have been disabled, following some time spent trying to
patch the version of sqlite. Hopefully we'll circle back to this with a solution.

The tox coverage was broadened to Python 3.10, too.